requirements.txt
htbuilder==0.5.0
torchvision==0.8.2
streamlit==0.84.1
torch==1.7.1
Pillow==8.3.1
